+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Soul urns should be fully functional (but cannot be crafted in survival due to lack of soul gem drops)
 - Ender crystal block effect was missing
+- Chaos coal blocks not working as fuel
 
 ## [3.0.3] - 2019-02-24
 ### Added